Can send the file, but has no filename and cannot output to standard out
We are successfully sending data to the server and saving it to a new
file, but the filename is being defaulted to '?' because the filename in
the first packet isn't being read. Also, the data that we need to print
out to the console for the TA to see cannot be printed right now.